IGAPP-1130: Add ExportEventButton
